Create SentenceClasi.py
A batch with three files: SentenceSplit.py, TrainModel.py, SentenceClasi.py
Using fasttext to do the training and classification. Result is not so satisfied, as whatever you input, fasttext will response the label.
Next step will try gensim, based on the similarity calculation. It will calculate the sentence similarity, from which it can decide if the sentence is really need further process.